Move Hive-related classes to mr.hive package
