+++ mr/src/test/java/org/apache/iceberg/mr/hive/serde/objectinspector/TestIcebergTimestampObjectInspector.java
@@ -17,47 +17,47 @@
  * under the License.
  */
 
-package org.apache.iceberg.mr.mapred.serde.objectinspector;
+package org.apache.iceberg.mr.hive.serde.objectinspector;
 
-import java.sql.Date;
-import java.time.LocalDate;
-import org.apache.hadoop.hive.serde2.io.DateWritable;
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
+import org.apache.hadoop.hive.serde2.io.TimestampWritable;
 import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspector;
 import org.apache.hadoop.hive.serde2.objectinspector.PrimitiveObjectInspector;
-import org.apache.hadoop.hive.serde2.objectinspector.primitive.DateObjectInspector;
+import org.apache.hadoop.hive.serde2.objectinspector.primitive.TimestampObjectInspector;
 import org.apache.hadoop.hive.serde2.typeinfo.TypeInfoFactory;
 import org.junit.Assert;
 import org.junit.Test;
 
-public class TestIcebergDateObjectInspector {
+public class TestIcebergTimestampObjectInspector {
 
   @Test
-  public void testIcebergDateObjectInspector() {
-    DateObjectInspector oi = IcebergDateObjectInspector.get();
+  public void testIcebergTimestampObjectInspector() {
+    TimestampObjectInspector oi = IcebergTimestampObjectInspector.get(false);
 
     Assert.assertEquals(ObjectInspector.Category.PRIMITIVE, oi.getCategory());
-    Assert.assertEquals(PrimitiveObjectInspector.PrimitiveCategory.DATE, oi.getPrimitiveCategory());
+    Assert.assertEquals(PrimitiveObjectInspector.PrimitiveCategory.TIMESTAMP, oi.getPrimitiveCategory());
 
-    Assert.assertEquals(TypeInfoFactory.dateTypeInfo, oi.getTypeInfo());
-    Assert.assertEquals(TypeInfoFactory.dateTypeInfo.getTypeName(), oi.getTypeName());
+    Assert.assertEquals(TypeInfoFactory.timestampTypeInfo, oi.getTypeInfo());
+    Assert.assertEquals(TypeInfoFactory.timestampTypeInfo.getTypeName(), oi.getTypeName());
 
-    Assert.assertEquals(Date.class, oi.getJavaPrimitiveClass());
-    Assert.assertEquals(DateWritable.class, oi.getPrimitiveWritableClass());
+    Assert.assertEquals(Timestamp.class, oi.getJavaPrimitiveClass());
+    Assert.assertEquals(TimestampWritable.class, oi.getPrimitiveWritableClass());
 
     Assert.assertNull(oi.copyObject(null));
     Assert.assertNull(oi.getPrimitiveJavaObject(null));
     Assert.assertNull(oi.getPrimitiveWritableObject(null));
 
-    LocalDate local = LocalDate.of(2020, 1, 1);
-    Date date = Date.valueOf("2020-01-01");
+    LocalDateTime local = LocalDateTime.of(2020, 1, 1, 0, 0);
+    Timestamp ts = Timestamp.valueOf("2020-01-01 00:00:00");
 
-    Assert.assertEquals(date, oi.getPrimitiveJavaObject(local));
-    Assert.assertEquals(new DateWritable(date), oi.getPrimitiveWritableObject(local));
+    Assert.assertEquals(ts, oi.getPrimitiveJavaObject(local));
+    Assert.assertEquals(new TimestampWritable(ts), oi.getPrimitiveWritableObject(local));
 
-    Date copy = (Date) oi.copyObject(date);
+    Timestamp copy = (Timestamp) oi.copyObject(ts);
 
-    Assert.assertEquals(date, copy);
-    Assert.assertNotSame(date, copy);
+    Assert.assertEquals(ts, copy);
+    Assert.assertNotSame(ts, copy);
 
     Assert.assertFalse(oi.preferWritable());
   }
